feat: add a profile card component; add hover & click effect; add closing tab on cards
